Rename includes to manipulate diagnostics.
Thanks to @mbrukman for making me think about it, also fix
Bazel build, we were missing a few header files in the CMake file.
