+++ bigtable/client/bigtable_client.bzl
@@ -15,6 +15,11 @@ bigtable_client_HDRS = [
     "internal/bulk_mutator.h",
     "internal/common_client.h",
     "internal/conjunction.h",
+    "internal/diagnostic_attributes_ignored.h",
+    "internal/diagnostic_pop.h",
+    "internal/diagnostic_push.h",
+    "internal/diagnostic_cxx14_extentions_ignored.h",
+    "internal/diagnostic_deprecated_as_warnings.h",
     "internal/encoder.h",
     "internal/endian.h",
     "internal/instance_admin.h",
