+++ bigtable/client/internal/diagnostic_attributes_ignored.h
