+++ bigtable/client/table_admin.h
@@ -22,8 +22,8 @@
 #include "bigtable/client/column_family.h"
 #include "bigtable/client/internal/port_platform.h"
 #include "bigtable/client/internal/table_admin.h"
-#include "bigtable/client/internal/wcxx14_extentions_ignored.h"
-#include "bigtable/client/internal/wattributes_ignored.h"
+#include "bigtable/client/internal/diagnostic_cxx14_extentions_ignored.h"
+#include "bigtable/client/internal/diagnostic_attributes_ignored.h"
 #include "bigtable/client/table_admin_strong_types.h"
 #include "bigtable/client/table_config.h"
 #include <memory>
